==> Quieter_Destinations_Group_Project/3. Recommended Destinations With Coordinates.ipynb <==
# Finding coordinates for each country #

### The OpenWeather API requires latitude and longitude coordinates in order to retrieve weather data ###
'Countries.csv' was used to provide these coordinates (via Google Dataset Publishing Language). However, in the first attempt of merge, multiple rows did not have values due to small mismatches like missing accents or rephrasing of name. This notebook shows the process of using fuzzy matching to obtain coordinates for those country names which are in fact the same place.

import pandas as pd 

from pathlib import Path

from thefuzz import process

from collections import Counter, defaultdict # For finding duplicates in the country name mapping without key error

data_dir = Path("project_data")

csv_file_path = data_dir / "full_sorted_tourism_with_recommendation.csv"

dest_df = pd.read_csv(csv_file_path, header=0)

# Checking data, 198 rows, Country, Mean total arrivals and Recommendations complete.

dest_df.info()

# Checking what data looks like

dest_df.head(25)

countries_path = data_dir / "countries.csv"

co_df = pd.read_csv(countries_path, header=0)  # dataset with latitude and longitude coordinates

# Checking data, 245 rows, no missing country names or coordinates.

co_df.info()

# Intrigued by country with no coordinates

co_df[co_df["latitude"].isna()]

# Makes sense as multiple places

# Checking what data looks like

co_df.head(25)

# In first attempt 36 rows had no values to be merged, 6 of which due to lack of capitalisation in And/Of.
# So using Title format for country names, to match for merge with detinations df.

co_df["name"]=co_df["name"].str.title() 

co_df.rename(columns={"country":"code", "name": "Country"}, inplace=True) # renaming columns so the two dataframes can be merged

co_df.head(30)

# In first attempt of merge, multiple rows did not have values due to small mismatches like missing accents or rephrasing of name.
# Using fuzzy matching to catch those country names which are in fact the same place.

dest_names = dest_df["Country"].unique().tolist()
co_names = co_df["Country"].unique().tolist()

dest_names # 198 country names

co_names # 245 country names, don"t need merge values for those extra not in dest df, so will map co to dest

# Using fuzzy mapping process to find matches for the names in coordinates data

fuzzy_mapping1 = {}
for co_name in co_df["Country"].unique():
    match, score = process.extractOne(co_name, dest_names)
    fuzzy_mapping1[co_name] = {"match": match, "score": score}

# Manually investigating non-exact matches

for co_name, info in fuzzy_mapping1.items():
    if info["score"] < 100:
        print(f"{co_name} -- {info["match"]} (Score: {info["score"]})")

# Not all of these non-matches are relevant.
# Only need to look at the destination names which do not have a perfect match.

matched_dest_names = {info["match"] for co_name, info in fuzzy_mapping1.items() if info["score"] == 100}


unmatched_dest_names = [name for name in dest_names if name not in matched_dest_names]

unmatched_dest_names

# Manually comparing this list with the <100 score matches, we can see that many countries have the correct fuzzy match.
# This covers: "Bolivia, Plurinational State Of", "Brunei Darussalam", "Cabo Verde", "Cote D´Ivoire", "Czech Republic (Czechia)",
# "Hong Kong, China", "Iran, Islamic Republic Of", "Micronesia, Federated States Of", "Moldova, Republic Of", "Myanmar", "North Macedonia",
# "Reunion", "Russian Federation", "Sao Tome And Principe", "Syrian Arab Republic", "Taiwan Province Of China","Tanzania, United Republic Of"
# "United States Of America" and "Viet Nam"


# Issues are: Congo (needs to not be republic and DRC separately, Curacao, Eswatini (is called Swaziland), Honk Kong, 
# Korea (want SK as NK no go), Lao, Macao, Sint Maarten, State Of Palestine, US Virgin Islands, 
# Turkey should be Turkiye not Turks and Cacos


# Manually inspecting the countries df to find the names for the above missing values.


cn = [n for n in co_names if "con" in n.lower()]
swa = [n for n in co_names if "swa" in n.lower()]
hong = [n for n in co_names if "hong" in n.lower()]
korea = [n for n in co_names if "korea" in n.lower()]
lao = [n for n in co_names if "lao" in n.lower()]
mac = [n for n in co_names if "mac" in n.lower()]
pal = [n for n in co_names if "palest" in n.lower()]
tai = [n for n in co_names if "taiwan" in n.lower()]
vn = [n for n in co_names if "virgin" in n.lower()]
cura = [n for n in co_names if "cura" in n.lower()]
maa = [n for n in co_names if "maarten" in n.lower()]

print(cn, swa, hong, korea, lao, mac, pal, tai, vn, cura, maa)

# Those with available matches are added to the manual overrides dictionary below.

manual_overrides = {
    "Congo [Republic]" : "Congo",
    "Congo [Drc]" : "Congo, Democratic Republic Of The",
    "Turkey" : "Türkiye",
    "Swaziland" : "Eswatini", 
    "Hong Kong" : "Hong Kong, China",
    "South Korea" : "Korea, Republic Of",
    "North Korea" : "Nowhere",
    "Laos" : "Lao People´S Democratic Republic",
    "Macau" : "Macao, China",
    "Palestinian Territories" : "State Of Palestine",
    "U.S. Virgin Islands" : "United States Virgin Islands" 
}

# No manual override for Sint Maarten and Curacao as neither is in the countries df under any other name.
# Therefore have searched for these manually to add in.
# CW Curacao lat: 12.1695 long: 68.990021
# MF Sint Maarten (Dutch Part) lat: 18.0528 long: -63.0425

# Mapping names of countries in the co_df to their fuzzy match, unless manually overriden

final_mapping = {}

for co_name in co_df["Country"].unique():
    if co_name in manual_overrides:
        final_mapping[co_name] = manual_overrides[co_name]
    else:
        final_mapping[co_name] = fuzzy_mapping1[co_name]["match"]

co_df["Country"].map(final_mapping).head(30)

# Noticed another problem here, Brunei appears twice and so would have two entries after merging.
# Realised this happens a few times as e.g. Algeria and Liberia both fuzzy matched to Algeria so the final mapping sent them both there. 
# Need to reverse engineer this to ensure that duplicates are sent "nowehere" to avoid giving multiple coordinates for one country.

# Detecting duplicate fuzzy matches, by counting how many times each dest name was mapped to

dest_counts = Counter(final_mapping.values())

# Find destinations that were matched more than once

duplicates = {dest for dest, count in dest_counts.items() if count > 1}
duplicates

# Need to choose only the 100 score fuzzy match for the duplicated destinations

match_scores = {
    (country, data["match"]): data["score"]
    for country, data in fuzzy_mapping1.items()
    if country not in manual_overrides
}

# Group which countries mapped to each destination

reverse_mapping = defaultdict(list)

for co_name, dest_name in final_mapping.items():
    reverse_mapping[dest_name].append(co_name)

reverse_mapping

# Reassign fuzzy duplicates only — preserving manual overrides

for dup in duplicates:
    matched_countries = reverse_mapping[dup] # list of countries that mapped there

    # Skip if only one match or if all are manual
    
    fuzzy_matches = [c for c in matched_countries if c not in manual_overrides] # removes manual overrides from matched 
    if len(fuzzy_matches) <= 1 or all(c in manual_overrides for c in matched_countries):
        continue 
    # Pick best fuzzy match (prefer score 100)

    # Default to None
    correct = None
    highest_score = -1
    highest_score_country = None
    
    for country in fuzzy_matches:
        score = match_scores.get((country, dup), 0) 
    
        if score == 100:
            correct = country
            break  # Prefer exact match, stop here
    
        if score > highest_score:
            highest_score = score
            highest_score_country = country # To look for next highest score through remaining
    
    if correct is None:
        correct = highest_score_country # Fall back to the best one if no 100
    
   
    for country in fuzzy_matches:
        if country != correct:
            final_mapping[country] = "Nowhere"

# Use the mapping to create column in co_df with names matching the dest_df

co_df["Destination Country"] = co_df["Country"].map(final_mapping)

# Can now merge the data-sets together

merge_df = pd.merge(dest_df, co_df[["Destination Country", "latitude", "longitude"]], left_on="Country", right_on="Destination Country",how="left")

merge_df

# Checking same number of rows in original dest_df and new merge_df

print(dest_df.shape)

print(merge_df.shape)

# No duplicates

# Checking for missing values

merge_df[merge_df["latitude"].isna()]

# Manually inserting the lat and long values for the final two destinations

merge_df.loc[merge_df["Country"] == "Curaçao", ["latitude", "longitude"]] = (12.1695, -68.990021)
merge_df.loc[merge_df["Country"] == "Sint Maarten (Dutch Part)", ["latitude", "longitude"]] = (18.0528, -63.0425)

merge_df[merge_df["latitude"].isna()] #Checking filled in

# Going forward we only need columns of tourism data with lat & long, so only keeping those.

cols = ["Unnamed: 0", "Country", "Arrivals", "Units", "2018", "2019", "2020", "2021", "2022", "Mean Total Arrivals", "Recommendation", "latitude", "longitude"]
full_recs_coords_df = merge_df[cols]

full_recs_coords_df.sample(30) # Checking looks right

# Double checking all 198 rows with no country, mean total arrivals, recommendation, lat or long missing. 

full_recs_coords_df.info() 

# # Saving as csv file ready for weather data to be added.

# complete_path = data_dir / "full_recommendations_with_coordinates.csv"

# full_recs_coords_df.to_csv(complete_path, index=False)
