==> Quieter_Destinations_Group_Project/4. Weather Data.ipynb <==
## Obtaining historic weather data using OpenWeather API ##

import requests
import pandas as pd
import numpy as np

# Using OpenWeather API to obtain weather data
base_url = "https://api.openweathermap.org/data/3.0/onecall/day_summary" 
api_key = "*****************************" # Enter your API key

# Function that requests and returns data from the OpenWeather API
def data(lat, lon, date):
    url = f"{base_url}?lat={lat}&lon={lon}&date={date}&appid={api_key}&units=metric"
    response = requests.get(url)

    if response.status_code == 200:
        requested_data = response.json()
        return requested_data
    else:
        print("Not found")

df = pd.read_csv('project_data/full_recommendations_with_coordinates.csv')

df.head(10)

### Trialling this function for coordinates in Crete, to ensure the function runs as expected ###

# # Requesting weather data from 2021-2024 for 6th June in Crete

# dates = ['2024-06-06', '2023-06-06', '2022-06-06', '2021-06-06']

# # API accepts latitude and longitude coordinates. These coordinates are for the centre of Crete.
# def crete_data(dates):
#     lat = 35.24012
#     lon = 24.80927
#     temps = [] 
#     total_rain = []

#     # Looping through each date to find the temperature and rainfall on that day.
#     for i in dates:
#         date = i
#         weather_data = data(lat, lon, date) # defining a variable for previous function.
#         temp_12pm = (weather_data['temperature']['afternoon']) # API docs specify 'afternoon' is 12pm temperature in that country.
#         rain = (weather_data['precipitation']['total'])

#         temps.append(temp_12pm)
#         total_rain.append(rain)

#     # Calculating mean average temperatures and rainfall across the four years
#     avg_temp = round(sum(temps) / len(temps))
#     avg_rainfall = round(sum(total_rain) / len(total_rain), 1)

#     return f" Avg_temp: {avg_temp}°C , Avg_rainfall: {avg_rainfall}mm"


# print(crete_data(dates))


# # Above function results:
# # temps = [28.89, 21.36, 27.84, 25.29]
# # total_rain = [1.19, 0.0, 0.0, 0.0]
# #  Avg_temp: 26°C , Avg_rainfall: 0.3mm


### Testing function with the first row of the dataframe ###

# dates = ['2024-06-06', '2023-06-06', '2022-06-06', '2021-06-06']

# # Requesting weather data from 2021-2024 on 6th June for countries in dataframe
# def trial_weather(dates):
#     temps = [] 
#     total_rain = []

#     # trialling with first index of dataframe
#     lat = df.loc[0,'latitude']
#     lon = df.loc[0,'longitude']
        
#     # Looping through each date to find the temperature and rainfall on that day.
#     for j in dates:
#         date = j
#         weather_data = data(lat, lon, date) # defining a variable for previous function.
#         temp_12pm = (weather_data['temperature']['afternoon']) # API docs specify 'afternoon' is 12pm temperature in that country.
#         rain = (weather_data['precipitation']['total'])

#         temps.append(temp_12pm)
#         total_rain.append(rain)

#     # Calculating mean average temperatures and rainfall across the four years
#     avg_temp = round(sum(temps) / len(temps))
#     avg_rainfall = round(sum(total_rain) / len(total_rain), 1)

#     return f" Avg_temp: {avg_temp}°C , Avg_rainfall: {avg_rainfall}mm"

# print(trial_weather(dates))


# Creating new columns to write average temperature and the average rainfall of each country on the 6th June
#every year from 2021-2024
df['avg_temp'] = 0
df['avg_rainfall'] = 0
df['avg_temp'] = df['avg_temp'].astype(float)
df['avg_rainfall'] = df['avg_rainfall'].astype(float)
df

# Filtering no-go destinations to see which API calls are not necessary to make
df[df['Recommendation'] == 'No Go']

Trialling skipping a row that has been determined as a no-go destination i.e. American Samoa, index 2

nogo_list = df.index[df['Recommendation'] == 'No Go'].tolist()

# Trialling weather function to see if it is successful in skipping index 2 (American Samoa)
dates = ['2024-06-06', '2023-06-06', '2022-06-06', '2021-06-06']

def trial_index_avg_weather(dates):
    temps = [] 
    total_rain = []

    # Ensuring index 2 gets looped through
    for i in range(3):
        lat = df.loc[i,'latitude']
        lon = df.loc[i,'longitude']

        # index 2 should be skipped in if statement
        if i in nogo_list:
            pass 
        else:
            for j in dates:
                date = j
                weather_data = data(lat, lon, date)
                temp_12pm = (weather_data['temperature']['afternoon']) 
                rain = (weather_data['precipitation']['total'])

                temps.append(temp_12pm)
                total_rain.append(rain)
            
                avg_temp = round(sum(temps) / len(temps))        
                avg_rainfall = round(sum(total_rain) / len(total_rain), 1)
        
            df.loc[i, 'avg_temp'] = avg_temp
            df.loc[i, 'avg_rainfall'] = avg_rainfall


trial_index_avg_weather(dates)

df

The function has skipped index 2, American Samoa and has left avg_temp and avg_rainfall at 0, reducing unneccessary API calls.

# Using function to iterate through entire dataframe, passing rows that are in the no go list
dates = ['2024-06-06', '2023-06-06', '2022-06-06', '2021-06-06']

def avg_weather(dates):
    temps = [] 
    total_rain = []

    for i in range(len(df)):
        lat = df.loc[i,'latitude']
        lon = df.loc[i,'longitude']

        if i in nogo_list:
            pass 
        else:
            for j in dates:
                date = j
                weather_data = data(lat, lon, date)
                temp_12pm = (weather_data['temperature']['afternoon']) 
                rain = (weather_data['precipitation']['total'])

                temps.append(temp_12pm)
                total_rain.append(rain)
            
                avg_temp = round(sum(temps) / len(temps))        
                avg_rainfall = round(sum(total_rain) / len(total_rain), 1)
        
            df.loc[i, 'avg_temp'] = avg_temp
            df.loc[i, 'avg_rainfall'] = avg_rainfall


avg_weather(dates)

df.sample(30)

Need to get weather data for Greece separately as weather data for new, quieter destinations need to be compared to Greece 

# Finding index in dataframe for Greece
gr_index = df.index[df['Country'] == 'Greece'].tolist()

# Creating a list for the index so this can be iterated through in weather function
old_dest = []
for i in gr_index:
    old_dest.append(i)

print(old_dest)

# Using function to get weather data for Greece
dates = ['2024-06-06', '2023-06-06', '2022-06-06', '2021-06-06']

def gr_weather(dates):
    temps = [] 
    total_rain = []

    for i in old_dest:
        lat = df.loc[i,'latitude']
        lon = df.loc[i,'longitude']
        for j in dates:
            date = j
            weather_data = data(lat, lon, date)
            temp_12pm = (weather_data['temperature']['afternoon']) 
            rain = (weather_data['precipitation']['total'])

            temps.append(temp_12pm)
            total_rain.append(rain)

            # Calculating mean averages of temperatures and rainfall
            avg_temp = round(sum(temps) / len(temps))        
            avg_rainfall = round(sum(total_rain) / len(total_rain), 1)
        
        df.loc[i, 'avg_temp'] = avg_temp
        df.loc[i, 'avg_rainfall'] = avg_rainfall


gr_weather(dates)

# Checking weather data for Greece has been added
print(df[df['Country'] == 'Greece'])

# Tidying up dataframe
df.rename(columns={'latitude': 'Latitude', 'longitude': 'Longitude', 'avg_temp': 'Avg Temp(°C)', 'avg_rainfall': 'Avg Rainfall(mm)'}, inplace = True)
df = df.drop(columns=['Unnamed: 0.1', 'Unnamed: 0'], errors='ignore')
df.head(15)

# # Writing results to a csv file
# df.to_csv('project_data/Complete_Weather_Data_For_Destinations.csv')
