==> Quieter_Destinations_Group_Project/5. The Three Recommended Destinations.ipynb <==
# Finding the top three destinations for our solo female traveller #

Using the Completed_Weather_Data_For_Destinations.csv file, filtered out the top destinations for our traveller by the weather (+- 2 degrees compared to Greece's average temp, and either equal to or lower than Greece's average rainfall)

import pandas as pd

df = pd.read_csv('project_data/Complete_Weather_Data_For_Destinations.csv')

df.head()

greece_df = df[df['Country'] == 'Greece']

greece_df.head()

greece_data = df[df['Country'] == 'Greece'].iloc[0]

greece_temp = greece_data['Avg Temp(°C)']
greece_rain = greece_data['Avg Rainfall(mm)']

# Filtering the countries based on temperature and rainfall in comparison to Greece.
top_destinations_df = df[
    (df['Avg Temp(°C)'] >= greece_temp - 2) &
    (df['Avg Temp(°C)'] <= greece_temp + 2) &
    (df['Avg Rainfall(mm)'] <= greece_rain)
]

top_destinations_df.head()

# Ordering the 5 countries from top_destinations_df from lowest total arrivals to highest

ordered_top_destinations = top_destinations_df.sort_values('Mean Total Arrivals').reset_index(drop=True)

ordered_top_destinations.head()

## The top three destinations are Aruba, Azerbaijan and Australia ##

This is based on finding destinations comparable to the climate of Greece (+- 2 degrees compared to Greece's average temperature, and either equal to or lower than Greece's average rainfall), whilst finding the quietest destinations based on mean total arrivals.
